Separates events card into its own component
