+++ web/src/features/channel/channel-page.module.scss
@@ -287,6 +287,6 @@ rowValue {
   }
 }
 .channelChartSettingsPopover {
-  min-width: 300px;
+  min-width: 150px;
   padding: 20px;
 }
